ticket:1618 - Petascope throws exception when disable_write_operations is enabled in petascope.properties
Summary:
In version 9.5, Petascope missed the check for this option in WCST parser.
It needs to add it back as same as previous versions which help users to protect imported coverages from being changed unexpectedly.

Test Plan: Tested with disable_write_operations=true when deleting coverage from WSClient and import coverage with WCST.

Reviewers: dmisev, vmerticariu

Differential Revision: http://codereview.rasdaman.org/D478