ticket:1620 - WMS requests are slow to handle in petascope
Summary:
Compare to version 9.4 which could serve multiple WMS requests from cache in milliseconds, version 9.5 takes few seconds to handle in PlantServer-2.
The problem comes from the call of gabarge collector in JVM. It should not call everytime and delay the time, a new request could be received in the controller. Only when it returns a big result, then it should remove the memory.
Also, fixed a problem with migration for WMS 3D layer, it should not throw exception, but just ignore this layer to migrate.

Test Plan: Tested with the mars_wgs84 tiff of PlanetServer-2.

Reviewers: dmisev, vmerticariu

Differential Revision: http://codereview.rasdaman.org/D479