ticket:1627 - Embedded SECORE needs user's configuration for secoredb path
Summary:
When running SECORE as embedded application, it needs a place to store secoredb and should not determine from db_update.path which may not exist
and later it will throw exception because it cannot find userdb collection.
With a valid folder path for secoredb, then, this problem will not happen when SECORE is set to embedded.

Test Plan: Tested manually.

Reviewers: dmisev

Differential Revision: http://codereview.rasdaman.org/D488