ticket:1631 - Petascope should not replace '-' with '_' for importing coverage id
Summary:
Petascope implicitly replaces '-' with '_' when creating rasdaman collection if a coverageid from wcst_import contains '-'.
However, as '-' cannot be part of coverageId/collectionName, so WCPS does not allow any queries to process on this kind of coverage.
This should throw exception from Petascope when it handles InsertCoverage request to be consistent.

Test Plan: Run test wcst_import for this test case.

Reviewers: dmisev, vmerticariu

Differential Revision: http://codereview.rasdaman.org/D491