ticket:1639 - Petascope hibernate shouldn't update a coverage's crs to database implicitly when it's crs is replaced from $ to configuration in petascope.properties
Summary:
A coverage's crs is stored in petascopedb as $SECORE_URL$ and later when petascop starts it will replace this placeholder with configuration from petascope.properties.
However, even this replacement doesn't save coverage to database, Hibernate still thinks this is a "dirty" update and wants to update this change to database automatically.
Then, later when secore is configured to a different endpoint (e.g: localhost:8081), this coverage cannot do anything with the new configured secore as the persisted crs is secore from old configuration (e.g: localhost:8080).

Test Plan: Tested with import coverage as localhost:8080 and later changes to localhost:8081 with DescribeCoverage.

Reviewers: vmerticariu, dmisev

Differential Revision: http://codereview.rasdaman.org/D495