ticket:1670 - RasdaPy_Signal to close thread properly to release connection to rasmgr/rasserver
Summary:
There is a bug in RasdaPy which it creates a thread to send a keep alive message to rasmgr/rasserver but this thread cannot be closed. Therefore,
cannot close db and connection to rasmgr/rasserver when the script is done and rasserver is occupied without being able to release (if rasql tries to connect later,
there will be exception 'no available rasservers').

Test Plan: Tested with running script to select, update, insert collections and afterwards, rasql can also connect to rasserver.

Reviewers: vmerticariu, dmisev

Differential Revision: http://codereview.rasdaman.org/D526