ticket:1644 - fix some errors with SECORE with BaseX version 8.6.7
Summary:
In BaseX 8.6.7, with incomplete URL request will return output string to split by "\n" instead of " " to build list of children.
With complete URL request, it will check the gml:identifier properly in userdb and gml collections. If a CRS definition contains a non-existing xlink:href CRS, it will return <empty/> element instead of throwing exception with gml_ version doesn't exist.
Also, the output from SECORE now will be consistent with the service.url prefix configured in secore.properties (before it was inconsistent and can be pointed to obsolete endpoint of local SECORE).

Test Plan:
It was tested with copying SECORE with BaseX version 7.3.1. Change service.url to different endpoint and check new output.
Run the test cases for SECORE in systemtest.

Reviewers: dmisev

Differential Revision: http://codereview.rasdaman.org/D534