ticket:1700 - Petascope accepts WCS GetCapabilities request without acceptversions parameter
Summary:
acceptversions is mandatory parameter in GetCapabilties? request, however a user might not know the supported versions until after seeing the Capabilities document.
Then, if a GetCapabilities request doesn't have this parameter, it uses the latest supported version (2.0.1) and considers it is a valid request.

Test Plan: Run the test for WCS GetCapabilities without acceptversions parameter.

Reviewers: vmerticariu, dmisev

Differential Revision: http://codereview.rasdaman.org/D546