ticket:1716 - unset http_proxy env variable in rasql
Summary: If http_proxy env variable is set in bash, then running a query using the rasql client from the terminal will direct the query to the proxy and the query will fail. To prevent these cases it is best to unset the http_proxy variable.

Test Plan: Manually tested.

Reviewers: vmerticariu

Reviewed By: vmerticariu

Differential Revision: http://codereview.rasdaman.org/D566