ticket:1727 - Log4j writes with datetime instead of only time to log files
Summary:
For log files from Log4j (e.g: petascope.log, secore.log), only time (HH:mm:ss) were logged. It is hard to know which date the requests were received and error happened.
The date and time should be added in the log files to know exactly when it happens.

Test Plan: It was tested with opengis.net for secore.log

Reviewers: dmisev, vmerticariu

Differential Revision: http://codereview.rasdaman.org/D575