ticket:1755 - Petascope WMS should do extend before scale for GetMap request's result
Summary:
e.g: sdom(coverage) is (0:860, 0:710) and GetMap request with grid bounds (700:1000, 500:800), only the intersection will be return (700:860, 500:710)
and if we bring this output to scale (e.g: scale(c[700:860, 500:710], [0:255,0:255])), this output will be stretched wrongly.
Therefore, it needs to use extend() for this out of the bbox case and after that, scale will keep the correct ratio.

Test Plan:
Tested with WebWorldWind with WMSLayer and it doesn't show the result out of the coverage's bounding box.
+ Before: https://imgur.com/a/s3ROADE
+ After: https://imgur.com/a/CX967ix

Differential Revision: http://codereview.rasdaman.org/D589