ticket:1713 - WSClient missing values in supported HTTP requests dropdown box from GetCoverage tab when selecting DescribeCoverage
Summary:
When selecting a coverage in WCS tab -> DescribeCoverage -> GetCoverage tab, the supported HTTP requests in the dropdown box is empty, because it is filled only when selecting a coverage in GetCoverage tab.

Test Plan: Test manually with DescribeCoverage and change to GetCoverage to see HTTP requests dropdown box has GET, POST requests.

Reviewers: dmisev, vmerticariu

Differential Revision: http://codereview.rasdaman.org/D590