ticket:1762 - Petascope delete created set/mdd/cell types from associated coverages when they are not used anymore
Summary: Petascope creates types for an importing coverage and these types could be reused for different coverage which shares same type structures.
However, if these types are not used and the coverage is deleted then these types also need to be removed as current there is an error about recreating type name when coverage is deleted, types are intact and a same coverageId is imported with different type structures (e.g: more bands).

Test Plan: Run the test in test WCST_Import

Reviewers: dmisev, vmerticariu

Differential Revision: http://codereview.rasdaman.org/D598