ticket:950 + ticket:1763 + ticket:1411 - support floating point null values, allow specifying null values in place in the form of `generalExp NULL VALUES [..]` and take into account null values on update. ticket:1699 - temporary conversion files are put in /tmp/rasdaman and *.aux.xml files are properly removed in the end.
Summary:
This patch adds support for floating point null values, e.g. `NULL VALUES [nan, 0.5:0.8]`.

A new table RAS_NULLVALUEPAIRS is added to RASBASE that stores the low/high null value pairs. An update_db.sh script migrates the old null values from RAS_DOMAINVALUES to RAS_NULLVALUEPAIRS.

Furthermore, it allows setting the null values of any MDD in a query in place. When this is used during update to specify null values for the source MDD, these do not overwrite the corresponding values in the target MDD.

Test Plan: Tests added to the nullvalues and manipulation tests in the systemtest.

Reviewers: vmerticariu

Reviewed By: vmerticariu

Differential Revision: http://codereview.rasdaman.org/D613