ticket:1787 - Petascope releases the lock from Liquibase when it starts if any
Summary:
This is a rare problem from Liquibase when it cannot finish properly and leaves a lock in petascopedb which blocks Petascope when Tomcat is restarted.
Therefore, when Petascope restarts, it will try to delete this lock and then, Liquibase will populate data normally.

Test Plan: This problem cannot be reproduced.

Reviewers: vmerticariu, dmisev

Differential Revision: http://codereview.rasdaman.org/D624