ticket:1811 - Error from WCPS should contain root exception if it exists
Summary:
Root exception (e.g: PetascopeException, SecoreException) in WCPS treenodes handlers are lost in some cases and it is hard to find from stacktrace.
Hence, root exception if it exists should be added to WCPSException to be shown in log of petascope for tracing to source error code.

Test Plan: Tested manually with verifying in the petascope's log.

Reviewers: vmerticariu, dmisev

Differential Revision: http://codereview.rasdaman.org/D650