ticket:1815 - WSClient - scroll to output console's bottom in WCS ProcessCoverages tab when a new WCPS query's result is added
Summary:
In WCS ProcessCoverages tab, when a new WCPS query's result is added from widgest (e.g: image>>, diagram>>), users have to scroll down
manually to see the result which is inconvenient. It should automatically do it to show the result to users.

Test Plan: Tested manually with running WCPS query when the scroll is not at the bottom of output console.

Reviewers: vmerticariu, dmisev

Differential Revision: http://codereview.rasdaman.org/D654