ticket:1872 - Petascope translate irregular coefficients to grid indices properly when coverage's origin is not the zero index in list of coefficients
Summary:
When coverage is imported with origin in the middle (coefficient zero in the middle of list of coefficients),
the tranlsation from geo bound to grid index has problem in this case as by default it can consider only origin is the first coefficient (index 0).

Test Plan:
Run test WCPS on a coverage imported with origin in the middle (insert a slice, then bottom slice, then top slice, sdom of time axis is (-1:1)).
The min date should go to the lowest index (-1) and the max date should go to the highest index (1).

Reviewers: dmisev, vmerticariu

Differential Revision: http://codereview.rasdaman.org/D717