ticket:1883 - WCSClient uses a specific GetCoverage button to download result
Summary:
Currently, it is ambiguous from GetCoverage's button when it needs 2 steps: step 1 select a coverage to show the core and extensions and step 2 to download result
from selected values. It is best to use the previous behavior to have a specific GetCoverage button to download result instead. Also, adjusted GUI of GetCoverage tab
to make it show the button clearly at the bottom as the main content has a vertical scroll not as before. The result looks like this: https://imgur.com/a/Uu2x2C9

Test Plan: Run test for test web interface (GetCoverage).

Reviewers: vmerticariu, dmisev

Differential Revision: http://codereview.rasdaman.org/D726