ticket:1892 - WMS add project() to each coverageExpression if request CRS is different from native CRS
Summary:
e.g: a WMS style is switch case $c > 0 and request CRS is EPSG:3857, nativeCRS is EPSG:4326
the generated rasql query should be project($c[0:100,0:200],..) > 0.

Also in generated Rasql of GetMap request, if scale and extend domains are same, so only scale domain is needed.

And reduceed the complexity of generated Rasql range constructor in WCPS.

Test Plan: Run test for WMS

Reviewers: vmerticariu, dmisev

Differential Revision: http://codereview.rasdaman.org/D735