ticket:1893 - Fix problem with WMS when zooming out of layer
Summary:
In WMS, the request BBox normally in EPSG:4326 and if layer's nativeCRS is different, it should not do anything to this original BBox.
Otherwise, miscalculated BBox in EPSG:4326 will result output in wrong location on the map.

Test Plan: Tested with coverage of Landsupport: https://imgur.com/a/BtAOefQ

Reviewers: vmerticariu, dmisev

Differential Revision: http://codereview.rasdaman.org/D738