ticket:1898 - WCPS subsetting by CRS:1 error when coverage imported with descending order in irregular axis
Summary:
Recently, a coverage can import slices by ascending (default) / descending order and this adds complexity to calculate Rasql grid bounds properly.
This patch is made to fix problem when subsettings in WCPS are Rasql grid indices directly via "CRS:1", and it should work also in this new type of importing data
to coverage when coefficient zero of irregular axis can be not the first element in the list of directPositions.

Test Plan: Tested in WCPS for various cases.

Reviewers: vmerticariu, dmisev

Differential Revision: http://codereview.rasdaman.org/D740