ticket:1896 - Add daemon functionality to wcst_import script
Summary:
Optionally, wcst_import is now able to run as a daemon process in the
background.

Starting wcst_import.sh on the same ingredients file while the daemon is
already running exists and the daemon keeps running.

Test Plan:
Errors are detected and handled properly.
The code commit was tested manually.

Reviewers: dmisev, vmerticariu

Reviewed By: dmisev

Differential Revision: http://codereview.rasdaman.org/D742