ticket:1894 - enhance WMS filling gaps when projection needed
Summary:
When requesting a WMS GetMap with BBox in different CRS from layer's native CRS, a projection is needed and the result can be rotated image
with gaps (nodata values) in its corners.
This problem will make the result of WMS GetMap in WMS clients like OpenLayers look not good with gaps between tiles.
Hence, this patch is implemented to try to fill the data in the gaps by selecting an extented BBox and then crop it by the original requested BBox in rasdaman.
The small gaps are still available and need to be enhanced more later.

Test Plan: Run test for WMS and see the enhanced result: https://imgur.com/a/ObcQzGm

Reviewers: vmerticariu, dmisev

Differential Revision: http://codereview.rasdaman.org/D765