ticket:1871 - Enhance custom Sentinel 2 recipe to analyze file immediately after reading it
Summary:
In custom Sentinel 2 recipe, it already created a Coverage object (GdalToCoverageConverter)
but it still needs to create axissubsets for this coverage, this is not redundant but it appears to show at misunderstanding place.
Hence, enhance the recipe to make it create coverage's axissubsets after fetching file immediately (1 Slice contains 1 axissubset)
which are parameters for UpdateCoverage request's parameters.

Test Plan: Tested but only has 1 Sentinel .SAFE file.

Reviewers: dmisev

Differential Revision: http://codereview.rasdaman.org/D790