ticket:1977 - Petascope axes and bands metadata are missing when updating coverage with local metadata
Summary:
Axes and bands metadata are removed when a next coverage slice requests to petascope to update coverage with its local metadata.
The problem is due to the way Jackson was configured to not display axes and bands metadata in coverage's global metadata.
Also, fix a problem with "\n" character in netCDF's history attribute in both petascope (Jackson allows to parse this character) and wcst_import (escape it properly).

Test Plan: Run test for test wcst_import (check awi test which contains axes, bands metadata and local metadata).

Reviewers: vmerticariu, dmisev

Differential Revision: http://codereview.rasdaman.org/D808