ticket:2069 - wsclient: Add loading icon for indicating running XmlHTTPRequest requests
Summary:
In wsclient, when an Ajax request is sent, a loading icon will show
to indicate about the progress and it will disappear when the request is done.

NOTE: Cannot detect WMS GetMap requests from WebWorldWind to Petascope as
the library uses a special mechanism (not normal XHR requests) to create Image elements dynamically
inside the library and add source GetMap URLs to them to load results.

Test Plan:
The result will look like this https://imgur.com/a/5zkj0fp when
an Ajax request is running.

Reviewers: vmerticariu, dmisev

Differential Revision: http://codereview.rasdaman.org/D866