ticket:2107 - WCPS uses colorPaletteTable extra parameter from query instead of the one in coverage's metadata
Summary:
In case coverage contains colorPaletteTable, it uses the value
from this one to add to rasql query for creating colorful image.
However, it does not regard to the corresponding extra parameter from WCPS queries
and is fixed to use the input colorPaletteTable if exists instead.

Test Plan: Run test for WCPS

Reviewers: dmisev, vmerticariu

Differential Revision: http://codereview.rasdaman.org/D904