ticket:2195 - Petascope shows valid coefficients for irregular axes with DescribeCoverage and encode in GML
Summary:
Petascope showed the invalid coefficients which are not usable by users as they are not adjusted properly with the lower geo bound
on the irregular axis. Also, fixed a problem with adding coefficients wrongly
instead of updating same coefficients when updating coverages.

Test Plan: Run test for test wcst_import and test wcs

Reviewers: vmerticariu, dmisev

Differential Revision: http://codereview.rasdaman.org/D967