ticket:2206 - wcst_import send WCS-T requests to petascope as POST
Summary:
Instead of sending a GML coverage file in /tmp folder to petascope,
wcst_import sends a POST request which contains the GML coverage in string
to insert/update a coverage. Petascope will not have a problem
to read the GML coverage file generated by wcst_import in /tmp folder if tomcat's security is enabled.

NOTE: it is tricky for handling the character & as wcst_import will not encode it, but curl will encode it for example.

Test Plan: Run test for wcst_import

Reviewers: vmerticariu, dmisev

Differential Revision: http://codereview.rasdaman.org/D979