ticket:2439 - Petascope subset error with approximate coefficients
Summary:
When a coefficient value has a large number of scale,
e.g: 4.66666666666666667, then it needs to check if it exists inside
the list of coefficients with a small epsilon number.
So, 4.66666666666666667 and 4.66666666666666666666666 will be equal
and the coefficient exists instead of throwing error.

Test Plan: Run the test WMS

Reviewers: dmisev

Differential Revision: https://codereview.rasdaman.com/D171